pyproject.toml
[project]
name = "stock-weather-agent"
version = "0.1.0"
description = "Workshop project: AI agent combining stock data and weather forecasts"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
bonus = [
    "pydantic-ai>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "assignments", "solutions", "bonus"]
